In jenv.fish, add flags to `set`s for safety
Use `-l` where I can to avoid overwriting any global variable called `cmd`, `arg`, or `script`. I don’t use it for the second `set arg` because it will have already been set, and I’m not sure if the `if` counts as a new scope for `-l` to create a new variable.

I also use `-gx` when setting `PATH` to make it global and export it, just to be sure.