Zulu provider now recognized aside from OpenJDK.
Reasons for distinction are nicely illustrated by Azul's CTO, Gil Tene:

https://groups.google.com/forum/#!topic/mechanical-sympathy/jQGahuzJKM4
https://news.ycombinator.com/item?id=8496003

And on Support forum for Zulu:
https://support.azul.com/hc/en-us/community/posts/200966724-Zulu-vs-OpenJDK-

Since I know of no other OpenJDK binaries which are CERTIFIED to be
TCK-compliant, seems like a good reason to have it.
Another is the fact, that SDKMAN installs it when you chose to install
Java.