Always add java bin directory to PATH in jenv-exec
The jenv exec command should prepend the path to the selected Java
version's bin directory to the PATH for the command being executed. The
command to be executed, however, is not necessarily a Java interpreter -
for example, the Maven plugin passes in /usr/local/bin/mvn. Therefore
the selected Java version's bin directory should not be determined from
the command, but from JAVA_HOME.