replace hardcoded path in jenv-macos-javahome
Replace a hardcoded call to `$HOME/.jenv/bin/jenv` with a call to the version of jenv that is in the `$PATH`.

This fixes an error when JENV is installed through homebrew.  (Fixes #230)

Alternatively this file could be edited during the brew installation to point to the path in Cellar.  I can close this and open a homebrew PR if that's the preferred way to do it.