Add architecture to the jenv alias format
Right now, jenv doesn't differentiate between JDKs for different
architectures, except for 32 bit vs 64 bit. For the vast majority of
systems, this is not a problem, since only one architecture is
supported. However, on the new M1 Macs, two 64 bit architectures are
supported: x86_64 and aarch64.

A user might want to be able to switch between two architectures of
JDKs. Current Jenv refuses to add the second JDK because it detects the
same version.

We can read the architecture by reading the `os.arch` system property.

This commit changes the jenv alias format from

  [provider][32|64]-[version]

to

  [provider]-[version]-[architecture]

For example, before, the full alias would be

  oracle64-17.0.2

and how it is

  oracle-17.0.2-x86_64

For M1 builds, it now shows

  orable-17.0.2-aarch64

This is a change to the alias format, so if you run `jenv add` with a
JDK that was already added to jenv, it will get re-added with the new
format.