Assume unknown shells are bash-compatible
Right now, jenv init only partially runs if it can't detect the shell.
Since bash-derived shells are probably the most common, let's just
assume that it is bash-compatible.

(I ran into this running from a x86 Docker container on an M1 Mac, where we
currently detect the shell as qemu-x86)