feat: handle system version better
Use java from the path with jenv shims and locations removed.
Set JAVA_HOME to be the linked java home from above, if it exists,
otherwise search for OS utility programs which can pinpoint the javahome
like libexec java_home or alternatives.
Fallback to whatever JAVA_HOME is currently set to if it can't otherwise
be resolved.