fix: prevent noclobber issues on /dev/null
This seems like the reason for the jenv-shim being left behind i.e.
rehash is run, the shim is created but we can't redirect to stderr
because of noclobber. This causes rehash to exit because of set -e but
does so before the trap is established to clean up the shim. Further
rehash executions fail because of the existence of the shim.

See https://github.com/rbenv/rbenv/issues/759#issuecomment-289326891